pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pyvndb"
version = "0.4.8"
description = "Vehicle description file handling for Python"
authors = ["Christoph Schueler <cpu12.gems@googlemail.com>"]
readme = "README.md"
license = "GPL-2.0-only"
repository = "https://www.github.com/Christoph2/pydbc"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "pydbc"}]
include = ["pydbc/cgen/templates/*.tmpl"]

[tool.poetry.dependencies]
python = ">=3.10"
antlr4-python3-runtime = "4.13.2"
chardet = ">=5.0.0"
mako = ">=1.2.0"
colorama = ">=0.4.6"
SQLAlchemy = ">=2.0.0"
typing-extensions = ">=4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-runner = ">=6.0.0"

[tool.poetry.group.docs.dependencies]
sphinxcontrib-napoleon = "*"
numpydoc = "*"
bumpversion = "*"

[tool.poetry.scripts]
vndb_importer = "pydbc.scripts.vndb_importer:main"
vndb_exporter = "pydbc.scripts.vndb_exporter:main"

[tool.pytest.ini_options]
addopts = "--verbose --tb=short -o junit_family=xunit2"
testpaths = ["pydbc/tests"]
